handlers/version.go
package handlers

import (
	"encoding/json"
	"log"
	"net/http"
)

// version returns a simple HTTP handler function which writes a JSON response.
func version(release string) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		info := struct {
			Release string `json:"release"`
		}{
			release,
		}

		body, err := json.Marshal(info)
		if err != nil {
			log.Printf("Could not encode info data: %v", err)
			http.Error(w, http.StatusText(http.StatusServiceUnavailable), http.StatusServiceUnavailable)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		w.Write(body)
	}
}
